requirements.txt
praw
mysql-connector-python
beautifulsoup4
pytesseract
Pillow
wordcloud
requests
python-dotenv